+++ composenav/src/main/java/com/adamkobus/compose/navigation/error/NavArgumentFormatInvalidError.kt
@@ -1,0 +1,3 @@
+package com.adamkobus.compose.navigation.error
+
+class NavArgumentFormatInvalidError(message: String, error: Throwable? = null) : RuntimeException(message, error)